Adding jury / contest deletion functionality

+++ src/main/java/com/photo_contest/services/ContestServiceImpl.java
@@ -95,6 +95,9 @@ public class ContestServiceImpl implements ContestService {
 
     @Override
     public void deleteContest(Long contestId) {
+        if (!contestRepository.existsById(contestId)) {
+            throw new EntityNotFoundException("Contest not found");
+        }
         contestRepository.deleteById(contestId);
     }
